Updated files from pre-commit check
