--- test/subset-band-name/reference_data/MOD02HKM.A2023309.2305.061.2023310073403_ref.hdf
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:bfb915be8434e7e0814960ff86393faff3fd3c0a05c962353da94e71a83a8324
-size 22229134
