+++ test/subset-band-name/reference_data/MOD02HKM.A2023309.2305.061.pscs_001729112207.hdf
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:187e211f5d5a38cfc3bb2ace30ba0c98682ef0006e59b3642a717c386679cf65
+size 31074789
